feat: Lyra theme - Wifi selection screen

+++ src/activities/network/WifiSelectionActivity.h
@@ -47,7 +47,7 @@ class WifiSelectionActivity final : public ActivityWithSubactivity {
   SemaphoreHandle_t renderingMutex = nullptr;
   bool updateRequired = false;
   WifiSelectionState state = WifiSelectionState::SCANNING;
-  int selectedNetworkIndex = 0;
+  size_t selectedNetworkIndex = 0;
   std::vector<WifiNetworkInfo> networks;
   const std::function<void(bool connected)> onComplete;
